Function cryptoil::sha::sha256::ops::digest_block_u32x4

pub fn digest_block_u32x4(state: (u32x4, u32x4), buf: &[u8]) -> (u32x4, u32x4)

Process a block with the SHA-2 SHA-256 algorithm.

Internally, this uses functions which resemble the new Intel SHA instruction sets, and so it's data locality properties may improve performance. However, to benefit the most from this implementation, replace these functions with x86 intrinsics to get a possible speed boost.